src/test/java/com/leetcode/strings/easy/ReverseStringII_541Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class ReverseStringII_541Test {
  private final ReverseStringII_541 solution = new ReverseStringII_541();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {

      return Stream.of(arguments("abcdefg", 2, "bacdfeg"));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void reverseStr(String str, int k, String expectedResult) {
    String actualResult = solution.reverseStr(str, k);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
